Slight refactor/rename around argument parsing
Rename 'PauseTargetProcess' to the more-correct 'ActivelyAttachToProcess'
in preparation for adding 'live' debug operations which do not pause but
do invade the process.

Use of IEnumerable for handling unparsed arguments irks me.
* Resharper complains about multiple enumeration of IEnumerable if we
  return the remaining args from ParseRemaining.
* In the usual case where we don't care about intermediate remaining args
  it's inelegant to follow 'var remaining = ...' with
  'remaining = <some function>(remaining)'.
In light of the above, pass an array by reference instead. The 'ref'
keyword makes it clear that the array may be replaced, while also
favouring the default use case of removing arguments as they are
processed. Modifying an IList is avoided because mutating the provided
collection instance is even less elegant than 'ref'.