Add detection and capture of Ctrl-C
Terminating a debugger attached to a process will bring that process down
with it. We should ensure that we have time to fully detach from a process
before we close, lest we bring that process down with us. Therefore it is
necessary that we gracefully handle eg. our hosting console session trying
to close.