Provide a 'safe' forced cast for use with generics
When an unconstrained generic type T has been tested at runtime to be of
a specific type, we must often cast via 'object' to satisfy the
compiler. But this causes the source type T to be ignored entirely, and
refactoring the surrounding code may cause the type of the source
variable to change, potentially invalidating our assumptions about the
validity of the cast.

To avoid accidental 'silent' changes to the source type, the ForceCast<,>
method requires that both the source and destination types be specified
explicitly.